feat(api-client): set request sidebar item context menu
