+++ packages/components/src/components/ScalarDropdown/ScalarDropdown.vue
@@ -3,25 +3,31 @@ import { Menu, MenuButton, MenuItems } from '@headlessui/vue'
 
 import { type FloatingOptions, ScalarFloating } from '../ScalarFloating'
 
-defineProps<Omit<FloatingOptions, 'middleware'>>()
+withDefaults(
+  defineProps<Omit<FloatingOptions, 'middleware'> & { static?: boolean }>(),
+  { static: false },
+)
 
 defineOptions({ inheritAttrs: false })
 </script>
 <template>
   <Menu v-slot="{ open }">
     <ScalarFloating
-      :isOpen="open ?? isOpen"
+      :isOpen="static ? true : open ?? isOpen"
       :placement="placement ?? 'bottom-start'"
       :resize="resize"
       :teleport="teleport">
-      <MenuButton as="template">
+      <MenuButton
+        v-if="!static"
+        as="template">
         <slot />
       </MenuButton>
       <template #floating="{ width }">
         <MenuItems
           class="relative flex w-56 flex-col p-0.75"
-          :style="{ width }"
-          v-bind="$attrs">
+          v-bind="$attrs"
+          :static="static"
+          :style="{ width }">
           <slot name="items" />
           <div
             class="absolute inset-0 -z-1 rounded bg-b-1 shadow-lg brightness-lifted" />
